requirements.txt
gradio==4.44.1
pandas
faiss-cpu
sentence-transformers

langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.17
langchain-text-splitters==0.2.4
langchain-groq==0.1.10

python-dotenv